Refactor My Shows feature to Continue Watching
* refactor: Rename presenter, state, and action classes from MyShows to ContinueWatching
* refactor: Update UI components and screens to reflect the Continue Watching nomenclature
* refactor: Update package names and directory structure to `com.thomaskioko.tvmaniac.continuewatching`
* refactor: Migrate and rename associated unit tests and Roborazzi screenshot baselines
* chore: Update module build configurations and project dependencies for the renamed feature modules
