+++ features/continue-watching/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/continuewatching/presenter/model/ContinueWatchingItem.kt
@@ -1,0 +1,16 @@
+package com.thomaskioko.tvmaniac.continuewatching.presenter.model
+
+public data class ContinueWatchingItem(
+    val traktId: Long,
+    val title: String,
+    val posterImageUrl: String? = null,
+    val status: String? = null,
+    val year: String? = null,
+    val seasonCount: Long = 0,
+    val episodeCount: Long = 0,
+    val episodesWatched: Long = 0,
+    val totalEpisodesTracked: Long = 0,
+    val watchProgress: Float = 0F,
+    val lastWatchedAt: Long? = null,
+    val nextEpisode: NextEpisodeItem? = null,
+)
