+++ features/continue-watching/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/continuewatching/presenter/model/EpisodeBadge.kt
@@ -1,0 +1,6 @@
+package com.thomaskioko.tvmaniac.continuewatching.presenter.model
+
+public enum class EpisodeBadge {
+    PREMIERE,
+    NEW,
+}
