+++ features/continue-watching/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/continuewatching/presenter/model/NextEpisodeItem.kt
@@ -1,0 +1,11 @@
+package com.thomaskioko.tvmaniac.continuewatching.presenter.model
+
+public data class NextEpisodeItem(
+    val episodeId: Long,
+    val episodeTitle: String,
+    val episodeNumberFormatted: String,
+    val seasonNumber: Long,
+    val episodeNumber: Long,
+    val stillPath: String?,
+    val firstAired: Long?,
+)
