+++ features/continue-watching/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/continuewatching/presenter/model/SectionedEpisodes.kt
@@ -1,0 +1,8 @@
+package com.thomaskioko.tvmaniac.continuewatching.presenter.model
+
+import kotlinx.collections.immutable.ImmutableList
+
+public data class SectionedEpisodes(
+    val watchNext: ImmutableList<UpNextEpisodeItem>,
+    val stale: ImmutableList<UpNextEpisodeItem>,
+)
