+++ features/continue-watching/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/continuewatching/presenter/model/SectionedItems.kt
@@ -1,0 +1,8 @@
+package com.thomaskioko.tvmaniac.continuewatching.presenter.model
+
+import kotlinx.collections.immutable.ImmutableList
+
+public data class SectionedItems(
+    val watchNext: ImmutableList<ContinueWatchingItem>,
+    val stale: ImmutableList<ContinueWatchingItem>,
+)
