+++ features/continue-watching/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/continuewatching/presenter/model/StartWatchingItem.kt
@@ -1,0 +1,8 @@
+package com.thomaskioko.tvmaniac.continuewatching.presenter.model
+
+public data class StartWatchingItem(
+    val traktId: Long,
+    val title: String,
+    val posterImageUrl: String?,
+    val year: String?,
+)
