+++ features/continue-watching/ui/src/test/screenshots/ContinueWatchingListGridView_light.png
