+++ features/continue-watching/ui/src/test/screenshots/ContinueWatchingListView_dark.png
