+++ features/continue-watching/ui/src/test/screenshots/ContinueWatchingListView_light.png
