+++ features/continue-watching/ui/src/test/screenshots/EmptyInProgress_dark.png
