+++ features/continue-watching/ui/src/test/screenshots/EmptySearchResult_dark.png
