+++ features/continue-watching/ui/src/test/screenshots/EmptySearchResult_light.png
