+++ features/continue-watching/ui/src/test/screenshots/ErrorLoadingShows_light.png
