+++ features/continue-watching/ui/src/test/screenshots/SearchActiveResult_dark.png
