+++ features/continue-watching/ui/src/test/screenshots/SearchActiveResult_light.png
