+++ features/continue-watching/ui/src/test/screenshots/WatchNextOnly_light.png
