--- features/my-shows/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/myshows/presenter/model/MyShowsItem.kt
@@ -1,16 +1,0 @@
-package com.thomaskioko.tvmaniac.myshows.presenter.model
-
-public data class MyShowsItem(
-    val traktId: Long,
-    val title: String,
-    val posterImageUrl: String? = null,
-    val status: String? = null,
-    val year: String? = null,
-    val seasonCount: Long = 0,
-    val episodeCount: Long = 0,
-    val episodesWatched: Long = 0,
-    val totalEpisodesTracked: Long = 0,
-    val watchProgress: Float = 0F,
-    val lastWatchedAt: Long? = null,
-    val nextEpisode: NextEpisodeItem? = null,
-)
