--- features/my-shows/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/myshows/presenter/model/NextEpisodeItem.kt
@@ -1,11 +1,0 @@
-package com.thomaskioko.tvmaniac.myshows.presenter.model
-
-public data class NextEpisodeItem(
-    val episodeId: Long,
-    val episodeTitle: String,
-    val episodeNumberFormatted: String,
-    val seasonNumber: Long,
-    val episodeNumber: Long,
-    val stillPath: String?,
-    val firstAired: Long?,
-)
