--- features/my-shows/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/myshows/presenter/model/SectionedEpisodes.kt
@@ -1,8 +1,0 @@
-package com.thomaskioko.tvmaniac.myshows.presenter.model
-
-import kotlinx.collections.immutable.ImmutableList
-
-public data class SectionedEpisodes(
-    val watchNext: ImmutableList<UpNextEpisodeItem>,
-    val stale: ImmutableList<UpNextEpisodeItem>,
-)
