--- features/my-shows/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/myshows/presenter/model/StartWatchingItem.kt
@@ -1,8 +1,0 @@
-package com.thomaskioko.tvmaniac.myshows.presenter.model
-
-public data class StartWatchingItem(
-    val traktId: Long,
-    val title: String,
-    val posterImageUrl: String?,
-    val year: String?,
-)
