--- features/my-shows/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/myshows/presenter/model/UpNextEpisodeItem.kt
@@ -1,19 +1,0 @@
-package com.thomaskioko.tvmaniac.myshows.presenter.model
-
-public data class UpNextEpisodeItem(
-    val showTraktId: Long,
-    val showName: String,
-    val showPoster: String?,
-    val episodeId: Long,
-    val episodeTitle: String,
-    val episodeNumberFormatted: String,
-    val seasonId: Long,
-    val seasonNumber: Long,
-    val episodeNumber: Long,
-    val runtime: String?,
-    val stillImage: String?,
-    val overview: String,
-    val badge: EpisodeBadge? = null,
-    val remainingEpisodes: Int = 0,
-    val lastWatchedAt: Long? = null,
-)
