--- features/my-shows/ui/src/main/kotlin/com/thomaskioko/tvmaniac/myshows/ui/component/Searchbar.kt
@@ -1,110 +1,0 @@
-package com.thomaskioko.tvmaniac.myshows.ui.component
-
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Clear
-import androidx.compose.material.icons.outlined.Search
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.TextRange
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.unit.dp
-import com.thomaskioko.tvmaniac.i18n.MR.strings.cd_clear
-import com.thomaskioko.tvmaniac.i18n.resolve
-
-@Composable
-internal fun Searchbar(
-    query: String,
-    hint: String,
-    onQueryChanged: (String) -> Unit,
-    onCloseClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val context = LocalContext.current
-    val focusRequester = remember { FocusRequester() }
-    var textFieldValue by remember(query) {
-        mutableStateOf(TextFieldValue(query, TextRange(query.length)))
-    }
-
-    LaunchedEffect(Unit) {
-        focusRequester.requestFocus()
-    }
-
-    OutlinedTextField(
-        modifier = modifier
-            .padding(end = 8.dp)
-            .fillMaxWidth()
-            .focusRequester(focusRequester),
-        value = textFieldValue,
-        onValueChange = { newValue ->
-            textFieldValue = newValue
-            onQueryChanged(newValue.text)
-        },
-        placeholder = {
-            Text(
-                text = hint,
-                style = MaterialTheme.typography.bodyMedium.copy(
-                    color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.6f),
-                ),
-            )
-        },
-        singleLine = true,
-        maxLines = 1,
-        textStyle = MaterialTheme.typography.bodyMedium,
-        keyboardOptions = KeyboardOptions(
-            keyboardType = KeyboardType.Text,
-            imeAction = ImeAction.Search,
-        ),
-        leadingIcon = {
-            Icon(
-                imageVector = Icons.Outlined.Search,
-                contentDescription = null,
-                tint = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.7f),
-            )
-        },
-        trailingIcon = {
-            IconButton(
-                onClick = {
-                    if (textFieldValue.text.isNotEmpty()) {
-                        textFieldValue = TextFieldValue()
-                        onQueryChanged("")
-                    } else {
-                        onCloseClick()
-                    }
-                },
-            ) {
-                Icon(
-                    imageVector = Icons.Filled.Clear,
-                    contentDescription = cd_clear.resolve(context),
-                    tint = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.7f),
-                )
-            }
-        },
-        shape = MaterialTheme.shapes.medium,
-        colors = TextFieldDefaults.colors(
-            focusedIndicatorColor = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.7f),
-            unfocusedIndicatorColor = MaterialTheme.colorScheme.outline.copy(alpha = 0.3f),
-            focusedContainerColor = MaterialTheme.colorScheme.surface,
-            unfocusedContainerColor = MaterialTheme.colorScheme.surface,
-            cursorColor = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.7f),
-        ),
-    )
-}
